Refactor form code from ObjectContentView to ObjectDataForm
- Move form state management (formData, errors) to ObjectDataForm
- Move form validation and change handlers to ObjectDataForm
- Move object types query and form initialization to ObjectDataForm
- Create callback props (onFormChange, onValidationChange) for parent communication
- Update ObjectContentView to use new ObjectDataForm interface with forwardRef
- Maintain UDC integration for coordinating between form data and widgets
- Improve separation of concerns and component reusability
